Organized code into seperate files, removed unused code, removed unneeded packages from ABSA requirements
