--- ASBA prototype/requirements.txt
@@ -1,63 +1,0 @@
-﻿annotated-types==0.7.0
-blis==1.2.0
-catalogue==2.0.10
-certifi==2025.1.31
-charset-normalizer==3.4.1
-click==8.1.8
-cloudpathlib==0.21.0
-colorama==0.4.6
-confection==0.1.5
-cymem==2.0.11
-filelock==3.18.0
-fsspec==2025.3.0
-huggingface-hub==0.29.3
-idna==3.10
-Jinja2==3.1.6
-joblib==1.4.2
-langcodes==3.5.0
-language_data==1.3.0
-marisa-trie==1.2.1
-markdown-it-py==3.0.0
-MarkupSafe==3.0.2
-mdurl==0.1.2
-mpmath==1.3.0
-murmurhash==1.0.12
-networkx==3.4.2
-nltk==3.9.1
-numpy==2.2.4
-packaging==24.2
-pillow==11.1.0
-preshed==3.0.9
-protobuf==6.30.1
-pydantic==2.10.6
-pydantic_core==2.27.2
-Pygments==2.19.1
-PyYAML==6.0.2
-rake-nltk==1.0.6
-regex==2024.11.6
-requests==2.32.3
-rich==13.9.4
-safetensors==0.5.3
-scikit-learn==1.6.1
-scipy==1.15.2
-sentence-transformers==3.4.1
-sentencepiece==0.2.0
-setuptools==76.0.0
-shellingham==1.5.4
-smart-open==7.1.0
-spacy-legacy==3.0.12
-spacy-loggers==1.0.5
-srsly==2.5.1
-sympy==1.13.1
-thinc==8.3.4
-threadpoolctl==3.6.0
-tokenizers==0.21.1
-torch==2.6.0
-tqdm==4.67.1
-transformers==4.49.0
-typer==0.15.2
-typing_extensions==4.12.2
-urllib3==2.3.0
-wasabi==1.1.3
-weasel==0.4.1
-wrapt==1.17.2
